EXAMPLE COMMIT
This commit uses the stm32f4 bsp to test out how to get a const struct
used as a flash config into a fixed location. This is bogus for the
stm32 micro, but the cpu is the same arch as the kinetis k64 bsp
I'm going to try writing.

For some unkown reason not all the test cases will build with this
linker script. It's almost like the flash config object is not
getting compiled.