Two tests run on the dev board
hello & ticker have successfully run. Console output goes to
a memory buffer instead of the serial port.

This imports a bunch fsl sdk code, some of which was changed
slightly and annotated saying so.

Next step is making the console work. After that refactor things
around so the headers and source files are better organized.