fixups for mk64fn
Disable the watchdog in the boot0 hook. Otherwise when powering on
with debugger disconnected the wdt would time out because boot_card
is too late causing the processor to continually reset.

Update flash config to disable ezport and nmi. Comment out some
arm cmsis stuff that was causing duplicate symbols in imported code
from the fsl.